+++ sample/BundleSample/BundleSample.cpp
@@ -8,61 +8,9 @@
 
 #include <iostream>
 
-#define RETURN_IF_FAILED(a) \
-    {   HRESULT __hr = a;   \
-        if (FAILED(__hr))   \
-        {   return __hr; }  \
-    }
+#include "Helpers.hpp"
 
-// Stripped down ComPtr provided for those platforms that do not already have a ComPtr class.
-template <class T>
-class ComPtr
-{
-public:
-    // default ctor
-    ComPtr() = default;
-    ComPtr(T* ptr) : m_ptr(ptr) { InternalAddRef(); }
-
-    ~ComPtr() { InternalRelease(); }
-    inline T* operator->() const { return m_ptr; }
-    inline T* Get() const { return m_ptr; }
-
-    inline T** operator&()
-    {   InternalRelease();
-        return &m_ptr;
-    }
-
-protected:
-    T* m_ptr = nullptr;
-
-    inline void InternalAddRef() { if (m_ptr) { m_ptr->AddRef(); } }
-    inline void InternalRelease()
-    {
-        T* temp = m_ptr;
-        if (temp)
-        {   m_ptr = nullptr;
-            temp->Release();
-        }
-    }
-};
-
-// Or you can use what-ever allocator/deallocator is best for your platform...
-LPVOID STDMETHODCALLTYPE MyAllocate(SIZE_T cb)  { return std::malloc(cb); }
-void STDMETHODCALLTYPE MyFree(LPVOID pv)        { std::free(pv); }
-
-// Helper class to free string buffers obtained from the packaging APIs.
-template<typename T>
-class Text
-{
-public:
-    T** operator&() { return &content; }
-    ~Text() { Cleanup(); }
-    T* Get() { return content; }
-
-    T* content = nullptr;
-protected:
-    void Cleanup() { if (content) { MyFree(content); content = nullptr; } }
-};
+using namespace MsixSample::Helper;
 
 int Help()
 {
@@ -264,4 +212,3 @@ int main(int argc, char* argv[])
 
     return 0;
 }
-
